fix(indexing): harden signed pad position mapping

+++ crates/tenferro-cpu/src/indexing.rs
@@ -1440,9 +1440,9 @@ fn typed_pad_with_fill<T: Copy + Clone + PoolScalar>(
     for input_value in input.as_slice()? {
         let mut in_bounds = true;
         for axis in 0..input_shape.len() {
-            let out_pos = config.edge_padding_low[axis]
-                + input_idx[axis] as i64 * (config.interior_padding[axis] + 1);
-            if !(0..out_shape[axis] as i64).contains(&out_pos) {
+            let out_pos = i128::from(config.edge_padding_low[axis])
+                + input_idx[axis] as i128 * i128::from(config.interior_padding[axis] + 1);
+            if !(0..out_shape[axis] as i128).contains(&out_pos) {
                 in_bounds = false;
                 break;
             }
